backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
psycopg2-binary
python-dotenv
pydantic
python-multipart
esdk-obs-python
Pillow
motor
pymongo
aiosmtplib
PyJWT
python-decouple
pydantic[email]
bcrypt